Restructure transfer paths
